series of several types can now be created from lvm and added to chart

+++ CryptoCenter/Interfaces/IHasUnixTime.cs
@@ -8,6 +8,6 @@ namespace CryptoCenter.Interfaces
 {
     public interface IHasUnixTime
     {
-        double UnixTime { get; set; }
+        int UnixTime { get; set; }
     }
 }
